pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-birthdays"
version = "0.1.0"
description = "Telegram birthday reminder bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot[job-queue]>=21.0,<22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
birthday-bot = "birthday_bot.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
